+++ src/losses.cpp
@@ -7,7 +7,7 @@ double mse(std::vector<double> yTrue, std::vector<double> yPred) {
     }
 
     double sum = 0;
-    for (int i = 0; i < yTrue.size(); i++) {
+    for (size_t i = 0; i < yTrue.size(); i++) {
         sum += pow(yTrue[i] - yPred[i], 2);
     }
 
@@ -22,7 +22,7 @@ std::vector<double> msePrime(std::vector<double> yTrue,
     }
 
     std::vector<double> gradient(yTrue.size());
-    for (int i = 0; i < yTrue.size(); i++) {
+    for (size_t i = 0; i < yTrue.size(); i++) {
         gradient[i] = 2 * (yPred[i] - yTrue[i]) / yTrue.size();
     }
 
@@ -38,7 +38,7 @@ double binaryCrossEntropy(std::vector<double> yTrue,
 
     double epsilon = 0.00001;
     double sum = 0;
-    for (int i = 0; i < yTrue.size(); i++) {
+    for (size_t i = 0; i < yTrue.size(); i++) {
         double value1 = std::max(yPred[i], epsilon);
         double value2 = std::max(1 - yPred[i], epsilon);
         sum += yTrue[i] * log(value1) + (1 - yTrue[i]) * log(value2);
@@ -56,7 +56,7 @@ std::vector<double> binaryCrossEntropyPrime(std::vector<double> yTrue,
 
     double epsilon = 0.00001;
     std::vector<double> gradient(yTrue.size());
-    for (int i = 0; i < yTrue.size(); i++) {
+    for (size_t i = 0; i < yTrue.size(); i++) {
         double value1 = std::max(yPred[i], epsilon);
         gradient[i] = ((1 - yTrue[i]) / (1 - yPred[i]) - yTrue[i] / value1) /
                       yTrue.size();
